+++ sample-apps/react-native/dogfood/src/components/VideoWrapper.tsx
@@ -3,11 +3,10 @@ import {
   StreamVideo,
   StreamVideoClient,
 } from '@stream-io/video-react-native-sdk';
-import { STREAM_API_KEY } from 'react-native-dotenv';
 import { useAppGlobalStoreValue } from '../contexts/AppContext';
 import { createToken } from '../modules/helpers/createToken';
 import translations from '../translations';
-import { customSentryLogger } from '../utils/logger';
+import { STREAM_API_KEY } from '../../config';
 
 export const VideoWrapper = ({ children }: PropsWithChildren<{}>) => {
   const userId = useAppGlobalStoreValue((store) => store.userId);
@@ -31,7 +30,7 @@ export const VideoWrapper = ({ children }: PropsWithChildren<{}>) => {
       apiKey: STREAM_API_KEY,
       user,
       tokenProvider: async () => createToken({ user_id: user.id }),
-      options: { logger: customSentryLogger, logLevel: 'warn' },
+      options: { logLevel: 'warn' },
     });
     setVideoClient(_videoClient);